requirements.txt
alembic==0.9.7
aniso8601==2.0.0
astroid==1.6.1
Babel==2.5.3
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Cors==3.0.3
Flask-Excel==0.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
isort==4.3.2
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
lml==0.0.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
Naked==0.1.31
passlib==1.7.1
pycryptodome==3.4.11
pyexcel==0.5.7
pyexcel-io==0.5.6
pyexcel-webio==0.1.4
pylint==1.8.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.2
texttable==1.2.1
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
